+++ src/main/java/com/codoacodo/flysky/demo/model/entity/VueloEntity.java
@@ -21,7 +21,7 @@ public class VueloEntity {
     private Integer numeroVuelo;
 
     private Boolean disponible;
-    private Long capacidad;
+    private Integer capacidad;
     private String aerolinea;
     private LocalDateTime fechaHoraPartida;
     private LocalDateTime fechaHoraLlegada;
